requirements.txt
attrs==22.1.0
beautifulsoup4==4.11.1
certifi==2022.6.15
charset-normalizer==2.1.0
filelock==3.7.1
gdown==4.5.1
idna==3.3
iniconfig==1.1.1
numpy==1.23.1
packaging==21.3
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.9
PySocks==1.7.1
pytest==7.1.2
PyYAML==6.0
requests==2.28.1
six==1.16.0
soupsieve==2.3.2.post1
tomli==2.0.1
tqdm==4.64.0
urllib3==1.26.11
